Move upload routes to /legacy/ for upload.pypi.io
